+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+RIPT only cares about the offsets so if it find partitions' earliest and latest offsets are the same it marks them as empty
+
+Topic State Records schema should be compatible with old and new versions of RIPT, try to avoid running different versions simultaneously on same Kafka cluster.
+
+### Added
+
+- /api/empty for listing empty topics
+- Ability to show empty topics on Dashboard
+
+### Changed
+
+- Topic State Record schema now includes is_empty field,  
+
 ## v0.2.0 - 2026-04-23
 
 ### Added